feat: Add spark run option for performance
When spark.sql.shuffle.partitions is not set it defaults to 200.
Depending on the size of the data that is being shuffled, this can cause
issues because spark does not compress shuffled data unless the number
of partitions is greather than 2000.  We will set the option here to
raise the partition count to 2001 to ensure we always have compressed
shuffles.