feat: Support new versions of Visonex data layouts
Visonex has recently changed the data layout for their source tables.
They have removed columns from a couple of tables as well as added a new
table that contain some of the removed columns.  The changes are as
follow:
  - patientdata: Removed ResuscitationCode, AdvDirResuscitationCode,
  AdvDirPowerOfAtty, AdvDirLivingWill, AdvDirPatientDeclines, AdvRevDate
  - patientmasterscheduleheader: Removed patientidnumber2,
  patientstatus2
  - Added new table AdvanceDirective, which contains the removed columns
  from patientdata
    - This is not included in the scope of this PR as we do not yet have
    any good data yet for this table

We address these changes by keeping what's common in one file and only
having separate files for the changes in the two tables mentioned above.