feat: Clean Advance Directive table for Visonex
We have now received the Advance Directive table for visonex for the
november 2018 and december 2018.  We add them to the sql files for
processing and include them in the spark routine as a table to unload.
The V2 start date does not need to change as 2018-11-30 is the first
date we received the advance directive table.