Fix multiple layers of quotes
In Python, parameters to functions are passed by reference. The run_spark_script
function adds quotes around a string value and modifies a passed in list of values
in place. Multiple calls to run_spark_script with the same list result in multiple
layers of quotes around the string value. Using deepcopy ensures that a new copy
of the list of values is passed into the function rather than the modified, quoted
values.