chore: Add build number to python3 container name
To more safely allow concurrent builds of the `DataPipeline-SparkTesting-Python3` Jenkins job, this commit adds the Jenkins build number to the container name. This is a precaution for if multiple builds start up within a second of each other, as the timestamp used in the container name (`$NOW`) goes down to seconds, not milliseconds.
QA-173